requirements.txt
flatbuffers==2.0
numpy==1.21.4
olefile==0.46
onnxruntime==1.9.0
opencv-python==4.5.4.58
Pillow==8.4.0
pip==21.3.1
protobuf==3.19.1
setuptools==59.2.0
wheel==0.37.0
